add/(added customerDto and tableserviceJava) fix/modified(fixed and modified few repositories and controllers)

+++ backend/src/main/java/com/pants/backend/controller/ReservationController.java
@@ -218,7 +218,7 @@ public class ReservationController {
                     .body(new ErrorResponse(400, "Customer is required"));
         }
 
-        Long customerId = reservation.getCustomer().getId();
+        Integer customerId = reservation.getCustomer().getId();
 
         if (customerId == null) {
             return ResponseEntity
